DEV-1815: Migrate integrate-with-vue3/ examples to Vue SFC format
Convert all remaining integrate-with-vue3/ examples from the legacy
split-file format (.js + .html) to Vue SFCs (.vue, <script setup>).

- 15 new .vue files created across 10 guide pages
- Old .js, .html, and .css files removed
- .md embedding directives updated from --html/--js flags to single .vue reference
- CSS from vue3-vuex-example embedded in <style> block
