--- docs/content/guides/integrate-with-vue3/vue3-custom-context-menu-example/vue/example1.html
@@ -1,3 +1,0 @@
-<div id="example1">
-  <hot-table :settings="hotSettings"></hot-table>
-</div>
