+++ docs/content/guides/integrate-with-vue3/vue3-custom-editor-example/vue/example1.vue
@@ -1,9 +1,9 @@
-import { defineComponent } from 'vue';
+<script setup>
+import { ref } from 'vue';
 import { HotTable } from '@handsontable/vue3';
 import { TextEditor } from 'handsontable/editors/textEditor';
 import { registerAllModules } from 'handsontable/registry';
 
-// register Handsontable's modules
 registerAllModules();
 
 class CustomEditor extends TextEditor {
@@ -19,28 +19,24 @@ class CustomEditor extends TextEditor {
   }
 }
 
-const ExampleComponent = defineComponent({
-  data() {
-    return {
-      hotSettings: {
-        startRows: 5,
-        columns: [
-          {
-            editor: CustomEditor
-          }
-        ],
-        colHeaders: true,
-        colWidths: 200,
-        autoWrapRow: true,
-        autoWrapCol: true,
-        height: 'auto',
-        licenseKey: 'non-commercial-and-evaluation'
-      }
-    };
-  },
-  components: {
-    HotTable,
-  }
+const hotSettings = ref({
+  startRows: 5,
+  columns: [
+    {
+      editor: CustomEditor
+    }
+  ],
+  colHeaders: true,
+  colWidths: 200,
+  autoWrapRow: true,
+  autoWrapCol: true,
+  height: 'auto',
+  licenseKey: 'non-commercial-and-evaluation'
 });
+</script>
 
-export default ExampleComponent;
+<template>
+  <div id="example1">
+    <HotTable :settings="hotSettings" />
+  </div>
+</template>
