--- docs/content/guides/integrate-with-vue3/vue3-custom-id-class-style/vue/example1.html
@@ -1,3 +1,0 @@
-<div id="example1">
-  <hot-table :id="id" :class="className" :style="style" :settings="hotSettings"></hot-table>
-</div>
