--- docs/content/guides/integrate-with-vue3/vue3-custom-id-class-style/vue/example1.js
@@ -1,30 +1,0 @@
-import { defineComponent } from 'vue';
-import { HotTable } from '@handsontable/vue3';
-import { registerAllModules } from 'handsontable/registry';
-
-// register Handsontable's modules
-registerAllModules();
-
-const ExampleComponent = defineComponent({
-  data() {
-    return {
-      hotSettings: {
-        startRows: 5,
-        startCols: 5,
-        colHeaders: true,
-        stretchH: 'all',
-        autoWrapRow: true,
-        autoWrapCol: true,
-        licenseKey: 'non-commercial-and-evaluation'
-      },
-      id: 'my-custom-id',
-      className: 'my-custom-classname',
-      style: 'height: 142px; overflow: hidden; border: 1px solid red;'
-    };
-  },
-  components: {
-    HotTable,
-  }
-});
-
-export default ExampleComponent;
