--- docs/content/guides/integrate-with-vue3/vue3-custom-renderer-example/vue/example2.html
@@ -1,3 +1,0 @@
-<div id="example2">
-  <hot-table :settings="hotSettings"></hot-table>
-</div>
