+++ docs/content/guides/integrate-with-vue3/vue3-formulas-example/vue/example1.vue
@@ -1,0 +1,39 @@
+<script setup>
+import { ref, markRaw } from 'vue';
+import { HotTable } from '@handsontable/vue3';
+import { registerAllModules } from 'handsontable/registry';
+import { HyperFormula } from 'hyperformula';
+
+registerAllModules();
+
+const hfRaw = markRaw(
+  HyperFormula.buildEmpty({
+    licenseKey: 'internal-use-in-handsontable',
+  })
+);
+
+const hotSettings = ref({
+  data: [
+    ['4', '=IF(A1>4, "TRUE", "FALSE")', 'C1'],
+    ['2', 'B2', '=A1+A2'],
+    ['3', 'B3', 'C3'],
+    ['4', 'B4', 'C4'],
+    ['5', 'B5', 'C5'],
+    ['6', 'B6', 'C6'],
+  ],
+  colHeaders: true,
+  height: 'auto',
+  autoWrapRow: true,
+  autoWrapCol: true,
+  licenseKey: 'non-commercial-and-evaluation',
+  formulas: {
+    engine: hfRaw,
+  },
+});
+</script>
+
+<template>
+  <div id="example1">
+    <HotTable :settings="hotSettings" />
+  </div>
+</template>
