--- docs/content/guides/integrate-with-vue3/vue3-hot-column/vue/example1.html
@@ -1,8 +1,0 @@
-<div id="example1">
-  <hot-table :settings="hotSettings">
-    <hot-column title="First column header">
-    </hot-column>
-    <hot-column :settings="secondColumnSettings" read-only="true">
-    </hot-column>
-  </hot-table>
-</div>
