--- docs/content/guides/integrate-with-vue3/vue3-hot-column/vue/example3.html
@@ -1,8 +1,0 @@
-<div id="example3">
-  <hot-table :data="hotData" :settings="settings">
-    <hot-column title="Column A" :editor="customEditor">
-    </hot-column>
-    <hot-column title="Column B" :read-only="true">
-    </hot-column>
-  </hot-table>
-</div>
