--- docs/content/guides/integrate-with-vue3/vue3-hot-reference/vue/example1.html
@@ -1,9 +1,0 @@
-<div id="example1">
-  <div class="example-controls-container">
-    <div class="controls">
-      <button v-on:click="swapHotData">Load new data</button>
-    </div>
-  </div>
-
-  <hot-table ref="hotTableComponent" :settings="hotSettings"></hot-table><br/>
-</div>
