--- docs/content/guides/integrate-with-vue3/vue3-hot-reference/vue/example1.js
@@ -1,37 +1,0 @@
-import { defineComponent } from 'vue';
-import { HotTable } from '@handsontable/vue3';
-import { registerAllModules } from 'handsontable/registry';
-
-// register Handsontable's modules
-registerAllModules();
-
-const ExampleComponent = defineComponent({
-  data() {
-    return {
-      hotSettings: {
-        data: [
-          ['A1', 'B1', 'C1', 'D1'],
-          ['A2', 'B2', 'C2', 'D2'],
-          ['A3', 'B3', 'C3', 'D3'],
-          ['A4', 'B4', 'C4', 'D4'],
-        ],
-        colHeaders: true,
-        height: 'auto',
-        autoWrapRow: true,
-        autoWrapCol: true,
-        licenseKey: 'non-commercial-and-evaluation'
-      }
-    };
-  },
-  methods: {
-    swapHotData() {
-      // The Handsontable instance is stored under the `hotInstance` property of the wrapper component.
-      this.$refs.hotTableComponent.hotInstance.loadData([['new', 'data']]);
-    }
-  },
-  components: {
-    HotTable,
-  }
-});
-
-export default ExampleComponent;
