+++ docs/content/guides/integrate-with-vue3/vue3-hot-reference/vue3-hot-reference.md
@@ -27,10 +27,9 @@ The following example implements the `@handsontable/vue3`, showing how to refere
 
 [Find out which Vue 3 versions are supported](@/guides/integrate-with-vue3/vue3-installation/vue3-installation.md#vue-3-version-support)
 
-::: example #example1 :vue3 --html 1 --js 2
+::: example #example1 :vue3
 
-@[code](@/content/guides/integrate-with-vue3/vue3-hot-reference/vue/example1.html)
-@[code](@/content/guides/integrate-with-vue3/vue3-hot-reference/vue/example1.js)
+@[code](@/content/guides/integrate-with-vue3/vue3-hot-reference/vue/example1.vue)
 
 :::
