--- docs/content/guides/integrate-with-vue3/vue3-language-change-example/vue/example1.html
@@ -1,29 +1,0 @@
-<div id="example1" class="select-language">
-  <div class="example-controls-container">
-    <div class="controls">
-      <label>Select language of the context menu:</label>
-      <div class="theme-dropdown" ref="dropdown">
-        <button
-          class="theme-dropdown-trigger"
-          type="button"
-          aria-haspopup="listbox"
-          :aria-expanded="isOpen"
-          @click="toggleDropdown"
-        >
-          <span>{{ language }}</span>
-          <svg class="theme-dropdown-chevron" aria-hidden="true" width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M6 9l6 6l6 -6"/></svg>
-        </button>
-        <ul class="theme-dropdown-menu" role="listbox" v-if="isOpen">
-          <li
-            v-for="lang in languages"
-            :key="lang"
-            role="option"
-            :aria-selected="language === lang"
-            @click="selectLanguage(lang)"
-          >{{ lang }}</li>
-        </ul>
-      </div>
-    </div>
-  </div>
-  <hot-table :language="language" :settings="hotSettings"></hot-table>
-</div>
