+++ docs/content/guides/integrate-with-vue3/vue3-language-change-example/vue3-language-change-example.md
@@ -33,10 +33,9 @@ Note that the `language` property is bound to the component separately using `la
 
 [Find out which Vue 3 versions are supported](@/guides/integrate-with-vue3/vue3-installation/vue3-installation.md#vue-3-version-support)
 
-::: example #example1 :vue3-languages --html 1 --js 2
+::: example #example1 :vue3-languages
 
-@[code](@/content/guides/integrate-with-vue3/vue3-language-change-example/vue/example1.html)
-@[code](@/content/guides/integrate-with-vue3/vue3-language-change-example/vue/example1.js)
+@[code](@/content/guides/integrate-with-vue3/vue3-language-change-example/vue/example1.vue)
 
 :::
