--- docs/content/guides/integrate-with-vue3/vue3-setting-up-a-language/vue/example1.js
@@ -1,57 +1,0 @@
-import { defineComponent } from 'vue';
-import { HotTable, HotColumn } from '@handsontable/vue3';
-import numbro from 'numbro';
-import jaJP from 'numbro/languages/ja-JP';
-import trTR from 'numbro/languages/tr-TR';
-import { registerAllModules } from 'handsontable/registry';
-
-// register Handsontable's modules
-registerAllModules();
-
-// register the languages you need
-numbro.registerLanguage(jaJP);
-numbro.registerLanguage(trTR);
-
-const ExampleComponent = defineComponent({
-  data() {
-    return {
-      formatJP: {
-        pattern: '0,0.00 $',
-        culture: 'ja-JP',
-      },
-      formatTR: {
-        pattern: '0,0.00 $',
-        culture: 'tr-TR',
-      },
-      hotData: [
-        {
-          productName: 'Product A',
-          JP_price: 1.32,
-          TR_price: 100.56,
-        },
-        {
-          productName: 'Product B',
-          JP_price: 2.22,
-          TR_price: 453.5,
-        },
-        {
-          productName: 'Product C',
-          JP_price: 3.1,
-          TR_price: 678.1,
-        },
-      ],
-      settings: {
-        height: 'auto',
-        autoWrapRow: true,
-        autoWrapCol: true,
-        licenseKey: 'non-commercial-and-evaluation'
-      }
-    };
-  },
-  components: {
-    HotTable,
-    HotColumn,
-  }
-});
-
-export default ExampleComponent;
