+++ docs/content/guides/integrate-with-vue3/vue3-setting-up-a-language/vue3-setting-up-a-language.md
@@ -27,10 +27,9 @@ The following example shows a Handsontable instance with translations set up in 
 
 [Find out which Vue 3 versions are supported](@/guides/integrate-with-vue3/vue3-installation/vue3-installation.md#vue-3-version-support)
 
-::: example #example1 :vue3-numbro --html 1 --js 2
+::: example #example1 :vue3-numbro
 
-@[code](@/content/guides/integrate-with-vue3/vue3-setting-up-a-language/vue/example1.html)
-@[code](@/content/guides/integrate-with-vue3/vue3-setting-up-a-language/vue/example1.js)
+@[code](@/content/guides/integrate-with-vue3/vue3-setting-up-a-language/vue/example1.vue)
 
 :::
