--- docs/content/guides/integrate-with-vue3/vue3-vuex-example/vue/example1.css
@@ -1,52 +1,0 @@
-#vuex-preview {
-  margin-top: 0.75rem;
-}
-
-#vuex-preview strong {
-  display: block;
-  margin-bottom: 0.375rem;
-  color: var(--sl-color-gray-2, #555555);
-  font-family: var(--sl-font, Inter, system-ui, -apple-system, sans-serif);
-  font-size: var(--sl-text-xs, 0.75rem);
-}
-
-#vuex-preview pre {
-  height: 168px;
-  padding: 0.5rem 0.75rem;
-  overflow-y: auto;
-  font-size: var(--sl-text-xs, 0.75rem);
-  font-family: var(--sl-font-mono, ui-monospace, monospace);
-  line-height: 1.6;
-  border: 1px solid var(--sl-color-gray-5, #e0e0e0);
-  background: var(--sl-color-gray-7, #f5f5f5);
-  color: var(--sl-color-gray-2, #555555);
-  margin: 0;
-  border-radius: 0;
-}
-
-#vuex-preview pre .table-container {
-  margin-bottom: 0.375rem;
-}
-
-#vuex-preview pre table {
-  border-collapse: collapse;
-  margin: 0.25rem 0;
-  font-size: var(--sl-text-xs, 0.75rem);
-}
-
-#vuex-preview pre td {
-  padding: 0.125rem 0.5rem 0.125rem 0;
-  color: var(--sl-color-gray-2, #555555);
-  border: none;
-}
-
-#vuex-preview pre ul {
-  margin: 0.125rem 0 0;
-  padding-left: 1.25rem;
-  list-style: disc;
-}
-
-#vuex-preview pre li {
-  padding: 0.0625rem 0;
-  color: var(--sl-color-gray-2, #555555);
-}
