--- docs/content/guides/integrate-with-vue3/vue3-vuex-example/vue/example1.html
@@ -1,12 +1,0 @@
-<div id="example1">
-  <div class="example-controls-container">
-    <div class="controls">
-      <label><input v-on:click="toggleReadOnly" checked id="readOnlyCheck" type="checkbox" /> Toggle <code>readOnly</code> for the entire table</label>
-    </div>
-  </div>
-  <hot-table ref="wrapper" :settings="hotSettings"></hot-table>
-  <div id="vuex-preview">
-    <strong>Vuex store dump:</strong>
-    <pre></pre>
-  </div>
-</div>
