+++ docs/content/guides/integrate-with-vue3/vue3-vuex-example/vue3-vuex-example.md
@@ -27,11 +27,9 @@ The following example implements the `@handsontable/vue3` component with a [`rea
 
 [Find out which Vue 3 versions are supported](@/guides/integrate-with-vue3/vue3-installation/vue3-installation.md#vue-3-version-support)
 
-::: example #example1 :vue3-vuex --html 1 --js 2 --css 3
+::: example #example1 :vue3-vuex
 
-@[code](@/content/guides/integrate-with-vue3/vue3-vuex-example/vue/example1.html)
-@[code](@/content/guides/integrate-with-vue3/vue3-vuex-example/vue/example1.js)
-@[code](@/content/guides/integrate-with-vue3/vue3-vuex-example/vue/example1.css)
+@[code](@/content/guides/integrate-with-vue3/vue3-vuex-example/vue/example1.vue)
 
 :::
